fix(parse-document): address code review feedback
- Remove unused pytest import from test_exceptions.py
- Rename abstract → abstract_text in DocumentMetadata
- Auto-derive full_markdown from pages via model_validator
- Change MinerUTimeoutError timeout type from int to float
- Add tests for full_markdown auto-derivation

+++ backend/src/core/ingest_and_digitize_data/parse_document/exceptions.py
@@ -17,7 +17,7 @@ class MinerUAPIError(ParseDocumentError):
 class MinerUTimeoutError(ParseDocumentError):
     """MinerU API request timed out."""
 
-    def __init__(self, timeout: int):
+    def __init__(self, timeout: float):
         self.timeout = timeout
         super().__init__(f"MinerU API timed out after {timeout}s")
